fix(domain): apply the second Codex review of #15
- Assessment carries the global low-confidence ceiling alongside the
  floor (level), so the renderer never has to recompute the minimum
  over verdicts on its own.
- GitActivity::medianFilesChanged is now ?float: an even-sized PR sample
  can legitimately yield a fractional median (e.g. 2 and 3 -> 2.5).
  SizeThresholds::bandForFiles() takes a float accordingly, with a
  boundary test pinning 2.5 to band M.
- Evidence's pointer is now a structural Pointer(file, field, value)
  value object instead of a non-empty string, so `new Evidence($claim,
  'git-activity.json')` — missing the field and value — can no longer
  be built. Pointer renders as "file > field = value" and refuses a
  blank file, field or value (MissingPointer, renamed to
  pointerFile/pointerField/pointerValue to avoid shadowing the
  non-readonly Exception::$file property). Tests updated accordingly.

+++ src/Domain/Profile/GitActivity.php
@@ -15,7 +15,7 @@ final readonly class GitActivity
     public function __construct(
         public ?Period $period,
         public ?int $pullRequestsTotal,
-        public ?int $medianFilesChanged,
+        public ?float $medianFilesChanged,
         public ?float $medianLinesChanged,
         public ?int $medianCorrectionCommitsAfterOpen,
         public ?int $mergedWithoutHumanEditAfterOpen,
